pyproject.toml
[project]
name = "sam-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "firecrawl>=4.3.1",
    "openai>=1.30.0",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.18.1",
    "tiktoken>=0.11.0",
]
